--- .circleci/config.yml
@@ -1,64 +1,0 @@
-version: 2.1
-
-jobs:
-  e2e: &test-run
-    parameters:
-      php-version:
-        type: string
-      symfony-version:
-        type: string
-    docker:
-      # node-browsers come with java pre-installed
-      - image: circleci/php:<< parameters.php-version >>-node-browsers
-    working_directory: ~/openzipkin
-    steps:
-      - checkout
-      - run: sudo apt install rsync
-      - run:
-          name: Install Composer
-          command: |
-            php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');"
-            php -r "if (hash_file('SHA384', 'composer-setup.php') === trim(file_get_contents('https://composer.github.io/installer.sig'))) { echo 'Installer verified'; } else { echo 'Installer invalid'; unlink('composer-setup.php'); } echo PHP_EOL;"
-            php composer-setup.php
-            php -r "unlink('composer-setup.php');"
-      - run:
-          name: Install Zipkin
-          command: curl -sSL https://zipkin.io/quickstart.sh | bash -s
-      - run:
-          name: Run Zipkin server
-          background: true
-          command: java -jar zipkin.jar
-      # start both processes in the background, but don't use composer as it dies
-      - run:
-          name: Build app
-          command: cd ./tests/E2E && make SYMFONY_VERSION=<< parameters.symfony-version >> LIBRARY_BRANCH=${CIRCLE_BRANCH} COMPOSER_RUNNER="php -d memory_limit=-1 /home/circleci/openzipkin/composer.phar" build
-      - run:
-          name: Run app
-          background: true
-          command: cd ./tests/E2E && make run-app
-      - run:
-          name: Run tests
-          command: cd ./tests/E2E && make test
-  e2e-6: *test-run
-
-workflows:
-  version: 2
-  all-tests:
-    jobs:
-      - e2e:
-          context:
-            - composer-auth # check https://app.circleci.com/settings/organization/github/jcchavezs/contexts
-          matrix:
-            parameters:
-              php-version: ["7.1", "7.2", "7.3", "7.4"]
-              symfony-version: ["4.4", "5.0"]
-            exclude:
-              - php-version: "7.1"
-                symfony-version: "5.0"
-      - e2e-6:
-          context:
-            - composer-auth # check https://app.circleci.com/settings/organization/github/jcchavezs/contexts
-          matrix:
-            parameters:
-              php-version: ["8.0", "8.1"]
-              symfony-version: ["6.0"]
