+++ tests/E2E/application/src/Sampler/CustomSampler.php
