+++ tests/Unit/KernelListenerTest.php
@@ -50,7 +50,7 @@ final class KernelListenerTest extends TestCase
         ];
     }
 
-    public function testSpanIsNotCreatedOnNonMasterRequest()
+    public function testSpanIsNotCreatedOnNonMainRequest()
     {
         /**
          * @var HttpServerTracing $httpServerTracing
@@ -60,12 +60,7 @@ final class KernelListenerTest extends TestCase
         $kernelListener = new KernelListener($httpServerTracing, RouteMapper::createAsNoop());
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(false);
-        } else {
-            $event->isMasterRequest()->willReturn(false);
-        }
-
+        $event->isMainRequest()->willReturn(false);
         $kernelListener->onKernelRequest($event->reveal());
 
         $this->assertNull(
@@ -96,11 +91,7 @@ final class KernelListenerTest extends TestCase
         ]);
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(true);
-        } else {
-            $event->isMasterRequest()->willReturn(true);
-        }
+        $event->isMainRequest()->willReturn(true);
         $event->getRequest()->willReturn($request);
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -130,11 +121,7 @@ final class KernelListenerTest extends TestCase
         $kernelListener = new KernelListener($httpServerTracing);
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(false);
-        } else {
-            $event->isMasterRequest()->willReturn(false);
-        }
+        $event->isMainRequest()->willReturn(false);
         $event->getRequest()->willReturn(new Request());
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -142,7 +129,7 @@ final class KernelListenerTest extends TestCase
         $exceptionEvent = new ExceptionEvent(
             $this->mockKernel(),
             new Request(),
-            HttpKernelInterface::SUB_REQUEST, // isMainRequest/isMasterRequest will be false
+            HttpKernelInterface::SUB_REQUEST, // isMainRequest will be false
             new Exception()
         );
 
@@ -164,11 +151,7 @@ final class KernelListenerTest extends TestCase
 
         $eventRequest = new Request();
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(true);
-        } else {
-            $event->isMasterRequest()->willReturn(true);
-        }
+        $event->isMainRequest()->willReturn(true);
         $event->getRequest()->willReturn($eventRequest);
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -176,7 +159,7 @@ final class KernelListenerTest extends TestCase
         $exceptionEvent = new ExceptionEvent(
             $this->mockKernel(),
             $eventRequest,
-            HttpKernelInterface::MASTER_REQUEST, // isMainRequest/isMasterRequest will be true
+            HttpKernelInterface::MAIN_REQUEST,
             new Exception(self::EXCEPTION_MESSAGE)
         );
 
@@ -201,11 +184,7 @@ final class KernelListenerTest extends TestCase
         $kernelListener = new KernelListener($httpServerTracing, RouteMapper::createAsNoop(), $logger);
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(false);
-        } else {
-            $event->isMasterRequest()->willReturn(false);
-        }
+        $event->isMainRequest()->willReturn(false);
         $event->getRequest()->willReturn(new Request());
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -251,11 +230,7 @@ final class KernelListenerTest extends TestCase
         ]);
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(true);
-        } else {
-            $event->isMasterRequest()->willReturn(true);
-        }
+        $event->isMainRequest()->willReturn(true);
         $event->getRequest()->willReturn($request);
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -263,7 +238,7 @@ final class KernelListenerTest extends TestCase
         $responseEvent = new ResponseEvent(
             $this->mockKernel(),
             $request,
-            KernelInterface::MASTER_REQUEST,
+            KernelInterface::MAIN_REQUEST,
             new Response('', $responseStatusCode)
         );
 
@@ -301,11 +276,7 @@ final class KernelListenerTest extends TestCase
         $request = new Request();
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(true);
-        } else {
-            $event->isMasterRequest()->willReturn(true);
-        }
+        $event->isMainRequest()->willReturn(true);
         $event->getRequest()->willReturn($request);
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -313,7 +284,7 @@ final class KernelListenerTest extends TestCase
         $responseEvent = new ResponseEvent(
             $this->mockKernel(),
             $request,
-            KernelInterface::MASTER_REQUEST,
+            KernelInterface::MAIN_REQUEST,
             new Response()
         );
 
@@ -346,11 +317,7 @@ final class KernelListenerTest extends TestCase
         ]);
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(true);
-        } else {
-            $event->isMasterRequest()->willReturn(true);
-        }
+        $event->isMainRequest()->willReturn(true);
         $event->getRequest()->willReturn($request);
 
         $kernelListener->onKernelRequest($event->reveal());
@@ -396,11 +363,8 @@ final class KernelListenerTest extends TestCase
         $request = new Request();
 
         $event = $this->prophesize(KernelEvent::class);
-        if (Kernel::MAJOR_VERSION >= 6) {
-            $event->isMainRequest()->willReturn(true);
-        } else {
-            $event->isMasterRequest()->willReturn(true);
-        }        $event->getRequest()->willReturn($request);
+        $event->isMainRequest()->willReturn(true);
+        $event->getRequest()->willReturn($request);
 
         $kernelListener->onKernelRequest($event->reveal());
